Authenticate GitHub calls via a ScopingHttpClient
Resolve the GitHub token once in the constructor and layer `auth_bearer`
onto the injected client with `ScopingHttpClient`, instead of rebuilding
the request options and re-reading the token on every paginated `tags()`
call.

Merge the token coverage into `TailwindVersionFinderTest`.
